workflows: move tasks and rename samples
